requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.1.31
click==8.1.8
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
docutils==0.21.2
executing==2.2.0
Faker==37.1.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jiter==0.9.0
jupyter_client==8.6.3
jupyter_core==5.7.2
marimo==0.12.8
Markdown==3.8
matplotlib-inline==0.1.7
narwhals==1.34.1
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.73.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycrdt==0.11.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pymdown-extensions==10.14.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
ruff==0.11.5
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.1
wcwidth==0.2.13
websockets==15.0.1
